Rawr3.BossHandler: - Added a DefaultMeleeAttack puller (and the UI to work with it) - Added a better Clone method for updating the Tab - Resorted some of the variables due to new practices - Changed the Picker's Primary layout to an accordion where you can select one group at a time and have it fill the tab - Added the Defensive group which covers the Boss's resistances (Frost, Physical, etc)
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C52060

+++ Rawr.Base/Bosses/BossList.cs
@@ -416,9 +416,9 @@ namespace Rawr {
             retboss.UseParryHaste = false;
             // Resistance
             foreach (ItemDamageType t in DamageTypes) {
-                value = passedList[0].Resistance(t);
+                value = (float)passedList[0].Resistance(t);
                 foreach (BossHandler boss in passedList) {
-                    value = Math.Min(value, boss.Resistance(t));
+                    value = (float)Math.Min(value, boss.Resistance(t));
                 }
                 retboss.Resistance(t, value);
             }
@@ -561,7 +561,7 @@ namespace Rawr {
             foreach (ItemDamageType t in DamageTypes) {
                 value = 0f;
                 foreach (BossHandler boss in passedList) {
-                    value += boss.Resistance(t);
+                    value += (float)boss.Resistance(t);
                 }
                 value /= passedList.Length;
                 retboss.Resistance(t, value);
@@ -737,7 +737,7 @@ namespace Rawr {
             foreach (ItemDamageType t in DamageTypes) {
                 value = 0f;
                 foreach (BossHandler boss in passedList) {
-                    value = Math.Max(value, boss.Resistance(t));
+                    value = (float)Math.Max(value, boss.Resistance(t));
                 }
                 retboss.Resistance(t, value);
             }
